pyproject.toml
[project]
name = "kusm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.2",
    "apscheduler>=3.11.2",
    "asyncio>=4.0.0",
    "datetime>=6.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

